Develop 유저 관련 코드 업데이트

See merge request s09-webmobile1-sub2/S09P12D205!5

+++ backend/src/main/java/com/ssafy/ssap/dto/RoomCreateDto.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 @Getter
 public class RoomCreateDto {
 
-    private Long userNo;
+    private int userNo;
     private String title;
     private long endHour;
     private long endMinute;
